Safari-609.1.11.1/LayoutTests/plugins/netscape-plugin-map-data-to-src-expected.txt

CONSOLE MESSAGE: line 13: PLUGIN: src: blah

This tests that an object tag with a data attribute gets a src attribute as well.
